requirements.txt
aniso8601==9.0.1
asttokens==2.0.8
backcall==0.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
decorator==5.1.1
dominate==2.7.0
dynaconf==3.1.9
executing==1.0.0
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Bootstrap==3.3.7.1
Flask-JWT-Extended==4.4.4
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
flask-shell-ipython==0.4.1
flask-simplelogin==0.1.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.3
idna==3.4
importlib-metadata==4.12.0
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
matplotlib-inline==0.1.6
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
PyJWT==2.5.0
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.41
SQLAlchemy-serializer==1.4.1
stack-data==0.5.0
traitlets==5.4.0
urllib3==1.26.12
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
